Extract mapping function to a separate utils file
